requirements.txt
ccxt==4.4.2
Flask==3.0.3
gspread==6.1.2
oauth2client==4.1.3
pandas==2.2.2
pycoingecko==3.1.0
python-dotenv==1.0.1
Requests==2.32.3
